requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
google-generativeai==0.8.1
matplotlib==3.9.2
seaborn==0.13.2
sqlalchemy==2.0.31
plotly==5.22.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0